restructure directories
