--- __init__.py
@@ -1,2 +1,0 @@
-# flake8: noqa: F401
-from .builder_factory import get_view_config_builder
