+++ src/builders/imaging_builders.py
