+++ test/fixtures/IMSViewConfBuilder/fake-entity.json
