+++ test/fixtures/ImagePyramidViewConfBuilder/fake-conf.json
