+++ test/fixtures/ImagePyramidViewConfBuilder/fake-entity.json
