+++ test/fixtures/NullViewConfBuilder/empty-conf.json
