+++ test/fixtures/NullViewConfBuilder/empty-entity.json
