+++ test/fixtures/RNASeqAnnDataZarrViewConfBuilder/fake-entity.json
