+++ test/fixtures/RNASeqViewConfBuilder/fake-entity.json
