+++ test/fixtures/SeqFISHViewConfBuilder/fake-conf.json
