+++ test/fixtures/SeqFISHViewConfBuilder/fake-entity.json
