+++ test/fixtures/StitchedCytokitSPRMViewConfBuilder/04e7385339167e541ad42a2636e18398-entity.json
