+++ test/fixtures/TiledSPRMViewConfBuilder/fake-entity.json
